Rename module css
