+++ components/avatar/styles.module.scss
