+++ components/footer/index.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import styles from "./footer.module.scss";
+import styles from "./styles.module.scss";
 
 export default function Footer({ name, github }) {
   return (
